Fixed package 0.2.0 setup relying on older README.
Summary:
Running `python setup.py install` or `pip install --upgrade facebookads` wasn't
working, because the package runs setup.py which expects an older version of
the README to be there.

Added a conditional check to see if the old README exists and to include that,
this way it will upgrade or install regardless of what version they're on.

Test Plan:
`pip install facebookads`
`pip install --upgrade facebookads`