ABSOLUTE_OCPM changed to  upper case as required by our api
Summary: BidType constant was lower-case and that impeded to create a local awareness adset on an example i'm creating

Test Plan:
execute the local_awareness.py file putting your adaccount in the 'env' variable. it should fail. apply this patch and it should work.

content of the script: https://phabricator.fb.com/P18123548