Fix bugs in fields handling, and make it consistent.
Summary:
The existing handling of fields is inconsistent in the three places it
appears: EdgeIterator, AbstractCrudObject.get_by_ids and
AbstractCrudObject.remote_read. The result varies between a string, a list,
None, or omitted from params.

This commit ensures that:

In all cases, params['fields'] will contain either a non-empty
string, or will be omitted entirely.

If fields is passed to a function, even an empty list, it will be
respected rather than falling back to get_default_read_fields. This
is accomplished by testing for is None rather than just a falsy
value.

Additionally this uses a more common Python idiom for duplicating a dict,
using dict(orig) rather than orig.copy().

https://github.com/facebook/facebook-python-ads-sdk/pull/22

Test Plan:
Run tests, we should also ensure that the following methods behave
as expected with:

1. fields defined
2. no fields defined
3. fields defined as empty []
4. fields defined via set_default_read_fields()

methods:
1. remote_read()
2. get_by_ids()
3. EdgeIterator method call (AdAccount.get_ad_sets())