Minor fixes to make sure Travis doesnt break
Summary:
I want to be able to retain the ability to run `tox` for all unit, integration,
and now docs tests. However, I don't want Travis to run integration nor docs
tests, because Travis won't have a config file properly filled out.

This diff adds a condition to the integration and docs tests that if the config
file is not in the repository it will gracefully exit as to not disturb tox.

Test Plan: Run tox