fix and rename add_users and remove_users
Summary: Primary thing to correct: both methods are broken because the key in `params` dict should be a string and not the arg from the function. I also changed the name because you add one user rather than multiple ones.

Test Plan:
Execute the code below. I didn't create a test case because @raven said he was working on a more thorough solution for that in some earlier diff.

  catalog = ProductCatalog(parent_id=business_id)
  catalog.update({ProductCatalog.Field.name: 'dpa bug catalog-1'})
  catalog.remote_create()
  print(catalog)
  ret = catalog.add_users(user_id,'ADMIN')
  print(ret)