Added a method to update a product on a ProductCatalog
Summary:
As of March 26, 2015, DPA feeds are downloaded daily. An advertiser may want to update a product is the middle of their cycle, for instance, if a product goes out of stock.
This call wasn't available in Python SDK, until now. More info about this API call in https://developers.facebook.com/docs/marketing-api/dynamic-product-ads/product-catalog?__mref=message_bubble#updateproduct

Test Plan:
The Python code should have the same behavior as the curl calls below (needs to call init before):
Parameters can be passed as named args or as a dict with key-value pairs of fields to be updated.

  catalog_id = '340252112833491'
  catalog = ProductCatalog(catalog_id)
  response = catalog.update_product('ID_1', price=300)

read price

  curl -G \
  -d 'appsecret_proof=b25cfd7373bde4e046b56f4de461a319839972f0a9b39b5f1fcc2bbfccf8c15c' \
  -d "access_token=CAAU5GZBVx5S4BAE34iDUGGl1eD8z2ZAESP5SpSLTlc500zKDVkhZBvX9aMptSqsbYG7JbgHaUUv4izt6t8CDkVkqRBUuYuXlUlGjdSIPfVMKtI8T4Ea0p6PTmZCkiFftOviAj6p2742qOWMHjXXpyYWbcY4IoIhhWagC47pPpCAZCqiOjQ35A" \
  https://graph.facebook.com/v2.3/catalog:340252112833491:SURfMQ==?fields=price

update price

  curl \
  -F 'price=5000' \
  -F 'appsecret_proof=b25cfd7373bde4e046b56f4de461a319839972f0a9b39b5f1fcc2bbfccf8c15c' \
  -F "access_token=CAAU5GZBVx5S4BAE34iDUGGl1eD8z2ZAESP5SpSLTlc500zKDVkhZBvX9aMptSqsbYG7JbgHaUUv4izt6t8CDkVkqRBUuYuXlUlGjdSIPfVMKtI8T4Ea0p6PTmZCkiFftOviAj6p2742qOWMHjXXpyYWbcY4IoIhhWagC47pPpCAZCqiOjQ35A" \
  https://graph.facebook.com/v2.3/catalog:340252112833491:SURfMQ==

read again to see that price was changed

  curl -G \
  -d 'appsecret_proof=b25cfd7373bde4e046b56f4de461a319839972f0a9b39b5f1fcc2bbfccf8c15c' \
  -d "access_token=CAAU5GZBVx5S4BAE34iDUGGl1eD8z2ZAESP5SpSLTlc500zKDVkhZBvX9aMptSqsbYG7JbgHaUUv4izt6t8CDkVkqRBUuYuXlUlGjdSIPfVMKtI8T4Ea0p6PTmZCkiFftOviAj6p2742qOWMHjXXpyYWbcY4IoIhhWagC47pPpCAZCqiOjQ35A" \
  https://graph.facebook.com/v2.3/catalog:340252112833491:SURfMQ==?fields=price