Fixing usage of default API all the time
Summary:
When using the SDK, one is allowed to pass an optional api parameter that defines which token is to be used with call from a given object in the constructor.

In case of classmethod calls, we need to implement this logic explicitly to avoid using the defaul API for call it shouldn't do. I browsed through the code looking for classmethods without an api optional argument and modified then. Happily, I just found one.

Test Plan:
`python -m facebookads.test.integration` and `python3 -m facebookads.test.integration`
Unit tests are not affected by this change, but feel free to run then as well.