Changed bootstrap script to be more robust
Summary:
The old bootstrap script was only intended to be used in interactive environment. I've changed it to be usable in test environment. This is also going to be used to make DocSmith examples simpler and with less cruft on example files.

Before, you needed insert the repository folder in PYTHONPATH, read config.json and call FacebookAdsApi.init:

  	import sys
  	import os
  	import json

  	this_dir = os.path.dirname(__file__)
  	repo_dir = os.path.join(this_dir, os.pardir, os.pardir) # this changes depengin on where your code resides
  	sys.path.insert(1, repo_dir)

  	from facebookads import FacebookAdsApi

  	auth_filename = os.path.join(repo_dir, 'config.json')
  	with open(auth_filename) as auth_file:
  	config = json.loads(auth_file.read())

  	FacebookAdsApi.init(
  	config['app_id'], config['app_secret'], config['access_token'])

Now you can just add to path, import bootstrap and call one function:

    	import sys
    	import os

    	this_dir = os.path.dirname(__file__)
    	repo_dir = os.path.join(this_dir, os.pardir, os.pardir)
    	sys.path.insert(1, repo_dir)

    	from facebookads import bootstrap
    	bootstrap.auth()

The code is a little simpler and, if already authenticated, auth() is a no-op.

Test Plan:
`python -i facebookads/boostrap.py` and on the REPL type

  from facebookads import api
  api.FacebookAdsApi.get_default_api()

Shouldn't be None.

Write a file such as:

  	import sys

  	repo_dir = '...'
  	sys.path.insert(1, repo_dir)

  	from facebookads import bootstrap
  	bootstrap.auth()

  	from facebookads import api
  	print(api.FacebookAdsApi.get_default_api())