app_secret dependency removed from FacebookAdsApi session creation
Summary: In cases when app_secret is not supplied it is valid to created the session with just the access_token alone. session.py is modified to cater to this case.

Test Plan: session.py is modify to create sessions with just the access_token alone when app_secret is not supplied. unit.py is modified with a test module added to test this particular case. The code is in the class SessionWithoutAppSecretTestCase