moving connections_id out of config
Summary:
https://phabricator.fb.com/P19837068
From the stacktrace in the paste it looks like the connections id in the the targeting spec was "<CONNECTIONS_ID>", so looks like it wasn't provided correctly trhough the config.json? In any case, I'll move the connections_id as a constant outside of the docblock.

Test Plan: run script