added DPA price update example
Summary: Commiting an example I used for live coding in an event in March 31st in São Paulo, Brazil. It randomly updates prices of products in real time.

Test Plan: Take a look at the README and check the code. In order to run, you'll need a Product Catalog created using the example feed provided.