get_insights_async bug fix
Summary:
This diff is associated with the PR: https://github.com/facebook/facebook-php-business-sdk/pull/517

**context:**
When `get_insights_async()` is called, it will check each field in the felds list belongs to AdReportRun which should be part of AdsInsights class.
**solution:**
Moved the fields to params to avoid the warning when `get_insights_async()` is called.

Reviewed By: jingping2015

Differential Revision: D18871811

fbshipit-source-id: 41c177ee