Error when event pushes
Facebook API does not accept an event without the "upload_tag" or "upload_id" so I changed the example to reflect that.  I do however still get a warning saying "data might not be compatible.  Expect list<string>; got <class 'list'>".   I am not sure what this error is or if it is hindering the API from accepting the event.