Updated typechecker.py to handle <list<map<string, string>>> correctly (#658)
Summary:
Fixed issue with not parsing <list<map correctly.

I first noticed this when getting products underneath a product catalog. One of the items in the request consisted of a <list<list<map<string, string>>> value_type. This is currently not handled when this is parsed in the existing code. In the existing code, sub_type_value tries to access an index of the subtypes list that does not exist. I noticed some similar code in the function get_typed_value which I've used in the fix.

Pull Request resolved: https://github.com/facebook/facebook-python-business-sdk/pull/658

Reviewed By: mengxuanzhangz

Differential Revision: D52914559

Pulled By: stcheng

fbshipit-source-id: 2db42d1ebfd4aab3047da857ef79ade66b80c4ef