Add `urllib3` to requirements for Retry support
This is needed since the addition of a Retry strategy which is provided by the `urllib3` library. This supports both PY2 (if we're really honestly still supporting this) and PY3 and was added since v1.9.0, so just pinning this to that version just in case someone is using an ancient (11+ years old) version of urllib3.