2T ThirdPartyAccount - Python SDK
Summary:
In order to simplify the usages for the two tier business model, we are creating SDK classes that abstract away complexity of setup in an effort to reduce onboarding friction. This contains both the classes as well as examples on how to use them.

The functionalities covers:
- creating a child business and associated ad account
- creating and sharing a line of credit with the child business
- fetching a particular child business account
- fetching the list of child business accounts
- getting the Facebook page of a child business
- token management is abstracted away as there are 3 tokens involved

This diff adds in the functionality for Python only.

Reference diff: D73810049

Differential Revision: D74041672

fbshipit-source-id: 9e97df3008d0fe1cf17d16c457e5d35cf50f038c